+++ massiv/tests/Data/Massiv/Core/IndexSpec.hs
@@ -38,6 +38,10 @@ instance (Index ix, Arbitrary ix) => Arbitrary (Sz ix) where
 instance (Index ix, Arbitrary ix) => Arbitrary (SzZ ix) where
   arbitrary = SzZ <$> liftIndex abs <$> arbitrary
 
+instance (Index ix, Arbitrary ix) => Arbitrary (Stride ix) where
+  arbitrary = do
+    Positive (Small x) <- arbitrary
+    Stride <$> liftIndex ((+1) . (`mod` x)) <$> arbitrary
 
 instance (Index ix, Arbitrary ix) => Arbitrary (SzIx ix) where
   arbitrary = do
@@ -85,6 +89,20 @@ instance CoArbitrary Ix4 where
 
 instance CoArbitrary Ix5 where
   coarbitrary (i :> ix) = coarbitrary i . coarbitrary ix
+
+
+-- instance Arbitrary Ix2 where
+--   arbitrary = (:.) <$> arbitrary <*> arbitrary
+
+-- instance (Arbitrary (Ix (n - 1))) => Arbitrary (IxN n) where
+--   arbitrary = (:>) <$> arbitrary <*> arbitrary
+
+-- instance CoArbitrary Ix2 where
+--   coarbitrary (i :. j) = coarbitrary i . coarbitrary j
+
+-- instance CoArbitrary (Ix (n - 1)) => CoArbitrary (IxN n) where
+--   coarbitrary (i :> ix) = coarbitrary i . coarbitrary ix
+
 
 
 prop_IsSafeIx :: Index ix => proxy ix -> SzIx ix -> Bool